// src/types/ThemeColors.ts
export interface ThemeColors {
    background: string;
    background2: string;
    text: string;
    primary: string;
    secondary: string;
    color1: string;
    color2: string;
    color3: string;
    textInputBackground: string;
    textInputBorder: string;
    textInputText: string;
    buttonBackground: string;
    buttonText: string;
    errorText: string;
    cardBackground: string;
    oddCardBackground: string;
    evenCardBackground: string;
    filtersBackground: string;
    tabBackground: string;
    tabText: string;
    biometricBox: string;
    biometricText: string;
  }
  